Munge the time value labels some
I'm not happy with the x-axis labelling at all, but this makes it a
little better.  The code is still pretty gross, but time is
unfortunately not as accomodating as the y-axis numeric values are.
It's hard to come up with natural looking time labels, especially if
you can't just round to the nearest day -- and even that gets tricky
with daylight savings time.